requirements.txt
dash
psycopg2-binary
pandas
plotly
